move repl to main to allow testing
