+++ include/ka/scheme/eval.hpp
@@ -1,27 +1,12 @@
 #pragma once
 
-#include "ka/scheme/fwd.hpp"
+#include "ka/scheme/types.hpp"
 
-#include <functional>
 #include <iosfwd>
-#include <map>
-#include <memory>
-#include <string>
 #include <variant>
 #include <vector>
 
 namespace ka::scheme {
-
-using Number = long double;
-using Symbol = std::string;
-using List = std::vector<ExprPtr>;
-using Function = std::function<ExprPtr(const List&)>;
-using Env = std::map<Symbol, ExprPtr>;
-struct Lambda {
-    std::vector<Symbol> params;
-    ExprPtr body;
-    Env closure;
-};
 using ExpressionTypes = std::variant<Symbol, Number, List, Function, Lambda>;
 
 struct Expr : public ExpressionTypes {
@@ -42,5 +27,4 @@ struct ExprOverloads : Ts... { using Ts::operator()...; };
 ExprPtr eval(ExprPtr&, Env&);
 ExprPtr eval_list(List& list, Env& env);
 std::ostream& print_expr(std::ostream& cout, ExprPtr expr);
-
 }  // namespace ka::scheme
