--- include/ka/scheme/fwd.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <memory>
-
-namespace ka::scheme {
-
-struct Expr;
-using ExprPtr = std::shared_ptr<Expr>;
-
-}  // namespace ka::scheme
